requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
flask-limiter>=3.5.0
requests>=2.25.0
paho-mqtt>=2.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
